Workaround for dismissed notifications returning
This is my workaround for dismissed Android notifications returning when
another app opens it's Media notification.  I made the notification
persistent when paused (cannot be dismissed) and stopped the
notification when Stop is pressed.
This might not be a complete solution but after chasing this one around
for 2 days it was the best work around I could come up with and seems
logical from my perspective. If the media player is paused it is still
active and the notification should show, if the media player is stopped,
it is not and the notification should go away.