IOS Refactoring & Partial UWP Refactoring
I was able to get all of the IOS Projects cleaned up.
UWP is kicking out 1300+ errors, System.Object not defined etc. I am unfamiliar with UWP and what may be missing (or may be .Net Standard bugs) causing the inability to compile.  I cleaned up what I could find in the errors list, so it should be close once the System dependencies or whatever is resolved.