Change to fix crash setting IOS volume before player is initialized
This change initializes the player at AudioImplementation intialization to ensure the player is initialized prior to being accessed in the Volume ManagerImplementation as the auto initialize in the Player property is inaccessible to the VolumeManagerImplementation.

If you think there is a cleaner implementation of this let me know, it was the easiest fix I could think of.